Migrated graph writer into stand-alone utility
