+++ descartes_light/solvers/include/descartes_light/solvers/bgl/impl/utils.hpp
@@ -5,17 +5,12 @@ DESCARTES_IGNORE_WARNINGS_PUSH
 #include <fstream>
 DESCARTES_IGNORE_WARNINGS_POP
 
-#include <descartes_light/solvers/bgl/boost_graph_types.h>
-
-const static std::string FILLCOLOR_ATTR = "fillcolor";
-const static std::string STYLE_ATTR = "style";
-const static std::string LABEL_ATTR = "label";
-const static std::string RANKDIR_ATTR = "rankdir";
+#include <descartes_light/solvers/bgl/utils.h>
 
 namespace descartes_light
 {
 template <typename FloatType>
-inline SubGraph<FloatType> createDecoratedSubGraph(const BGLGraph<FloatType>& g)
+SubGraph<FloatType> createDecoratedSubGraph(const BGLGraph<FloatType>& g)
 {
   SubGraph<FloatType> main;
   boost::get_property(main, boost::graph_name) = "G";
@@ -95,39 +90,13 @@ inline SubGraph<FloatType> createDecoratedSubGraph(const BGLGraph<FloatType>& g)
 }
 
 template <typename FloatType>
-void BGLLadderGraphSolver<FloatType>::writeGraph(const std::string& filename) const
+void writeGraph(const std::string& filename, const BGLGraph<FloatType>& graph)
 {
   std::ofstream file(filename);
   if (!file.good())
     throw std::runtime_error("Failed to open file '" + filename + "'");
 
-  boost::write_graphviz(file, createDecoratedSubGraph(graph_));
-}
-
-template <typename FloatType>
-void BGLLadderGraphSolver<FloatType>::writeGraphWithPath(const std::string& filename) const
-{
-  std::ofstream file(filename);
-  if (!file.good())
-    throw std::runtime_error("Failed to open file '" + filename + "'");
-
-  SubGraph<FloatType> sg = createDecoratedSubGraph(graph_);
-
-  // Get the path through the graph
-  auto target = std::min_element(ladder_rungs_.back().begin(),
-                                 ladder_rungs_.back().end(),
-                                 [this](const VertexDesc<FloatType>& a, const VertexDesc<FloatType>& b) {
-                                   return graph_[a].distance < graph_[b].distance;
-                                 });
-  const std::vector<VertexDesc<FloatType>> path = reconstructPath(source_, *target);
-
-  // Colorize the path
-  for (const VertexDesc<FloatType>& v : path)
-  {
-    boost::get(boost::vertex_attribute, sg)[v][FILLCOLOR_ATTR] = "green";
-  }
-
-  boost::write_graphviz(file, sg);
+  boost::write_graphviz(file, createDecoratedSubGraph(graph));
 }
 
 }  // namespace descartes_light
